backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.2.3
Pillow==9.0.0
opencv-python-headless==4.5.5.64
pytesseract==0.3.9
numpy==1.22.0
spacy==3.2.0
google-generativeai==0.3.1
requests==2.27.1
ultralytics==8.0.0
gunicorn==20.1.0
huggingface-hub==0.19.0
torch>=1.7.0
tqdm==4.64.0
python-dotenv==1.0.0
doclayout-yolo==0.0.3